Convert Digest to a BigInt
It turns out that a Digest is really just a BigInt.  Right now it's a
sub-class, and I think I'm going to keep it that way.  The only
problem is that we'll be able to do arithmetic on BigInts and it
doesn't make a whole lot of sense to do arithmetic on Digests.