Remove BigInt(void)
This only existed because I wasn't properly constructing some of my
BigInts in a constructor, and is now removed.  The only reason I added
it in the first place was because I was too lazy to lookup the syntax,
but it's pretty unsafe so if I'm going to clean up code I might as
well do it correctly :).