Add some sanity checks to ModInt
The ModInt class is optimized such that it will only work when "data <
mod".  This is the only well-formed case (it's a modular field, so
other numbers simply don't exist), so it has to be checked for.

There's a small convience function where I allow numbers such that
"data < 2*mod" to exist.  This essentially relies on the fact that the
moduli are all very large in crypto algroithms.  Allowing for these
larger numbers means that we can accept raw RNG output (where "data >=
mod"), but disallowing everything else avoids a DOS (where a very
small modulus is used for a very large number).