Clean up the last commit
I simply moved the logic for mod from the ModInt constructor into the
BigInt % operator.  This fixes the problem because I needed to
implement this anyway.

Note that this is somewhat of a hack: I still only support "x % m"
when "x < 2m".  As long as the higheset bit is set in the mod then
this will never be a problem.  Also, note that the ModInt class
prevents this from ever being true during add/subtract.