Implement "BigInt % BigInt"
I generated a few random test vectors, and it seems one of them
consistantly fails while the others don't.  I'm not sure why this is,
I've looked at it for a bit and it doesn't seem any different.

The other 9 tests (4 random, 5 constructed) pass.  Hopefully that
means it's something simple.