Fix multiplication overflow bug
Essentially the problem is that, while a 64x64->128 multiply can't
overflow, it can overflow when we accumulate a whole bunch.  That
means I need a 64x64->256 (well, I use one -- it probably doesn't need
to be nearly that big) multiply-accumulate unit in the middle.

This fixes that mystery ModInt failure that I had before.