Functioning NIST "Point + Point"
This doesn't do point doubling, but it does at least work for point
addition.  It turns out the big problem involved NIST's definition of
a (which is different than Wikipedia, but the same as Certicom).

Additionally, there was a bug in the modular inversion code related to
it returning a BigInt instead of a ModInt -- that was stupid!