Run RISC-V tests on Spike
This is more of a test of the RISC-V toolchain than any actual test of
my code.  It does at least run things on RISC-V and report correct
answers, so that's good.

There are a few assorted changes related to the fact that we're using
newlib instead of glibc on RISC-V and newlib doesn't have some
functionality.