Remove ModInt::invert tests for non-prime moduli
It turns out the hardware algorithm we're using only handles prime
moduli for inversions.  This is fine for ECDSA because the moduli is
always prime, so I've gone ahead and removed everything with a
non-prime moduli (converting some of the tests and removing others).

Unfortunately, the Chisel code still doesn't seem to want to work for
me.  Hopefully it's just a bug in my test harness?